requirements.txt
fastapi
uvicorn
cohere
python-dotenv
requests
